# Conflicts:
#	smartervote-domain/src/main/java/ch/zhaw/smartervote/domain/ElectionProposalServiceImpl.java
#	smartervote-domain/src/main/java/ch/zhaw/smartervote/domain/algorithm/ElectionProposalAlgorithm.java
#	smartervote-domain/src/main/java/ch/zhaw/smartervote/domain/algorithm/QuestionAnswerMatcher.java
#	smartervote-domain/src/test/java/ch/zhaw/smartervote/domain/ElectionProposalServiceIntegrationTest.java
#	smartervote-interface/src/main/java/ch/zhaw/smartervote/contract/ElectionProposalService.java

+++ smartervote-domain/src/main/java/ch/zhaw/smartervote/domain/mapping/MapPersonalQuestion.java
@@ -32,7 +32,7 @@ public class MapPersonalQuestion {
             answer = new PersonalAnswerTO(entity.getPersonalQuestionAnswer().getText());
         }
 
-        return new PersonalQuestionTO(entity.getId(), entity.getText(), (int) entity.getUpvotes(), upvoted, answer);
+        return new PersonalQuestionTO(entity.getId(), entity.getText(), entity.getUpvotes(), upvoted, answer);
     }
 
     /**
